refactor: cleanup gdsii code
