+++ src/renderer/lib/gdsii/struct.js
