+++ src/renderer/lib/gdsii/testdata/1Kpolyg.gds
