+++ src/renderer/lib/gdsii/types.ts
@@ -1,0 +1,4 @@
+export type RecordToken = {
+  recordType: number
+  data: ArrayBuffer | number[] | string
+}
