+++ src/renderer/lib/gdsii/utils.ts
@@ -1,0 +1,8 @@
+export function buf2hex(buffer: ArrayBuffer) {
+  // buffer is an ArrayBuffer
+  return [...new Uint8Array(buffer)].map((x) => x.toString(16).padStart(2, '0')).join('')
+}
+
+export function isEmpty(obj) {
+  return Object.keys(obj).length === 0
+}
